ZNB20 file was now useless since everybody use ZNB20V2 instead. ZNB20V2 is renamed ZNB20. Think to update your create_intrument file.
